initial efforts to compile for Dotty, issue #9
- adds Dotty version 0.24.0-RC1 because Scala-Test is not yet available
  for 0.26.0-RC1
- fixing a number of missing empty-parens issues
- fixing missing do-while syntax
- adding a few return types
- some other syntax quirks
- update to ScalaTest 3.2.0
- found a potential issue in RetrySet (tailrec call with default values
  instead of actual values)
- fix some problems with `Function0` syntax

Sadly it does not compile yet, because the compiler does not implicitly
resolve any instance of `OptManifest`. No idea why?

+++ src/main/scala-2.13+/scala/concurrent/stm/skel/AtomicArrayBuilder.scala
@@ -93,7 +93,7 @@ object AtomicArrayBuilder {
   class ofByte extends IntBacked[Byte] {
     def addOne(elem: Byte): this.type = {
       ensureSpace()
-      elems(size) = elem
+      elems(size) = elem.toInt
       size += 1
       this
     }
@@ -107,7 +107,7 @@ object AtomicArrayBuilder {
   class ofShort extends IntBacked[Short] {
     def addOne(elem: Short): this.type = {
       ensureSpace()
-      elems(size) = elem
+      elems(size) = elem.toInt
       size += 1
       this
     }
@@ -121,7 +121,7 @@ object AtomicArrayBuilder {
   class ofChar extends IntBacked[Char] {
     def addOne(elem: Char): this.type = {
       ensureSpace()
-      elems(size) = elem
+      elems(size) = elem.toInt
       size += 1
       this
     }
